Move Card/UIPopupButton/ResetChangesModal to shared, delete dead component-utils.ts
Completes the theme-builder migration in ag-grid-docs: moves the last
generic pieces into ag-website-shared/src/components/theme-builder:

- Card - trivial styled div, only remaining consumer was UIPopupButton
- UIPopupButton - generic floating popup/dropdown button, now using
  the shared Card, the classnames package instead of grid's local
  combineClassNames, and a consolidated useClickAwayListener
- ResetChangesModal + Dialog.module.scss - the generic "are you sure
  you want to reset?" confirmation modal; content is theme-builder
  generic copy, not grid-specific, and the SCSS already only pulls
  from the shared design-system partial

Also extracts useClickAwayListener.ts as a real shared export instead
of leaving a private copy duplicated inside ColorPicker.tsx from an
earlier move.

Deletes component-utils.ts entirely - it was already fully dead
(memoWithSameType, useChangeHandler, useEffectWithCustomDependencies,
and its own local getThemeDefaultParams had no remaining consumers;
combineClassNames/useClickAwayListener were its only live exports,
both superseded by the move above).

Rewires GetTheme.tsx, GridConfigDropdown.tsx, and PresetSelector.tsx
onto the shared locations.

This closes out the grid-side migration - the full theme-builder tree
was swept again (not just editors/) and everything remaining is
either page-shell/composition-root wiring (ThemeBuilder.tsx,
RootContainer.tsx, the register*.ts files) or genuinely host-specific
content (presets, grid-config demo scaffolding, EmptyState,
BordersEditor's border list, ThemeImportExportDialog's code-gen).

Verified: tsc identical to baseline, eslint/prettier clean, 40/40
tests pass, and manually in the running dev server - the Import/Export
popup opens correctly, and making two changes then switching presets
correctly triggers the shared ResetChangesModal with working
Cancel/Reset.
