+++ documentation/ag-grid-docs/src/components/theme-builder/components/grid-config/GridConfigDropdown.tsx
@@ -1,10 +1,10 @@
 import { Checkbox } from '@ag-website-shared/components/theme-builder/Checkbox';
 import { Tooltip } from '@ag-website-shared/components/theme-builder/Tooltip';
+import { UIPopupButton } from '@ag-website-shared/components/theme-builder/UIPopupButton';
 import { titleCase } from '@ag-website-shared/theming/utils';
 import { WarningAltFilled } from '@carbon/icons-react';
 import styled from '@emotion/styled';
 
-import { UIPopupButton } from '../general/UIPopupButton';
 import { useGridConfigAtom } from './grid-config-atom';
 import { allConfigFields, incompatibleGridConfigProperties, productionConfigFields } from './grid-options';
