+++ external/ag-website-shared/src/components/theme-builder/UIPopupButton.tsx
@@ -9,11 +9,12 @@ import {
     size,
     useFloating,
 } from '@floating-ui/react';
+import classnames from 'classnames';
 import { type ReactNode, useEffect, useRef, useState } from 'react';
 import type React from 'react';
 
-import { combineClassNames, useClickAwayListener } from '../component-utils';
 import { Card } from './Card';
+import { useClickAwayListener } from './useClickAwayListener';
 
 export type UIPopupButtonProps = {
     dropdownContent: ReactNode | ((close: () => void) => ReactNode);
@@ -61,7 +62,6 @@ export const UIPopupButton = (props: UIPopupButtonProps) => {
 
     const prevShow = useRef(show);
     const onCloseRef = useRef(props.onClose);
-    // eslint-disable-next-line react-hooks/refs -- keeping ref in sync with latest callback
     onCloseRef.current = props.onClose;
     useEffect(() => {
         if (prevShow.current && !show) {
@@ -75,11 +75,10 @@ export const UIPopupButton = (props: UIPopupButtonProps) => {
     return (
         <>
             <Button
-                className={combineClassNames(
-                    props.className,
-                    show && 'is-dropdown-visible',
-                    `variant-${props.variant || 'primary'}`
-                )}
+                className={classnames(props.className, {
+                    'is-dropdown-visible': show,
+                    [`variant-${props.variant || 'primary'}`]: true,
+                })}
                 onClick={() => setShow(!show)}
                 ref={refs.setReference}
                 color="neutral"
@@ -89,7 +88,6 @@ export const UIPopupButton = (props: UIPopupButtonProps) => {
                 {props.endDecorator}
             </Button>
             {show && (
-                // eslint-disable-next-line react-hooks/refs -- floating-ui callback ref pattern
                 <DropdownArea ref={refs.setFloating} style={floatingStyles} className={props.dropdownClassName}>
                     <div className="dropdownWrapper">
                         {typeof props.dropdownContent === 'function'
